requirements.txt
mss>=9.0.0
Pillow>=10.0.0
imageio>=2.31.0
numpy>=1.24.0
pyinstaller>=5.13.0
